requirements.txt
# --- Core dependencies ---
langchain==0.3.4
langchain-community==0.3.3
langchain-google-genai==1.0.5
langchain-text-splitters==0.3.0

# --- Vector store & embeddings ---
chromadb==0.5.11
sentence-transformers==3.2.1

# --- PDF and text handling ---
pypdf==4.3.1

# --- Web UI ---
streamlit==1.39.0

# --- Environment variables ---
python-dotenv==1.0.1

# --- Utility packages ---
tqdm==4.66.5
numpy>=1.26.0